+++ Frontend/src/lib/images/index.ts
@@ -14,6 +14,8 @@ export {default as circuitPattern} from './404/circuit-pattern.svg';
 export {default as itCrowdGif} from './404/IT-Crowd.gif';
 
 export {default as satisfaction} from './homepage/satisfaction.png';
+export {default as dashboard} from './homepage/dashboard.png';
+
 export {default as cog} from './homepage/features/cog.svg';
 export {default as tag} from './homepage/features/tag.svg';
 export {default as shield} from './homepage/features/shield.svg';