+++ Frontend/tailwind.config.js
@@ -13,7 +13,7 @@ export default {
       boxShadow: {'8xl': '0 0 30px 30px rgba(0, 0, 0, 0.20)'},
       animation: {
         border: 'border 4s linear infinite',
-        shake: 'shake-up-bottom 6s ease-in-out infinite',
+        shake: 'shake 6s ease-in-out infinite',
       },
       keyframes: {
         border: {to: {'--border-angle': '360deg'}},